Add testing of bit_ceil

+++ include/boost/int128/bit.hpp
@@ -29,7 +29,7 @@ constexpr int bit_width(const uint128_t x) noexcept
 
 constexpr uint128_t bit_ceil(const uint128_t x) noexcept
 {
-    return x <= 1 ? 1 : static_cast<uint128_t>(1) << bit_width(x - 1);
+    return x <= 1 ? static_cast<uint128_t>(1) : static_cast<uint128_t>(1) << bit_width(x - 1);
 }
 
 } // namespace int128
